+++ include/libft/ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: epacheco <epacheco@student.42sp.org.       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/11 13:35:49 by epacheco          #+#    #+#             */
-/*   Updated: 2021/09/11 13:35:50 by epacheco         ###   ########.fr       */
+/*   Updated: 2021/12/11 18:23:06 by epacheco         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@ char	*ft_strjoin(char const *s1, char const *s2)
 	char	*str;
 	int		size;
 
-	size = ft_strlen(s1) + ft_strlen(s2);
-	str = (char *)malloc(sizeof(char) * (size + 1));
+	size = (int)(ft_strlen(s1) + ft_strlen(s2));
+	str = (char *)malloc(sizeof(char) * ((unsigned long)size + 1));
 	if (!str || !s1 || !s2)
 		return (NULL);
 	ft_memcpy(str, s1, ft_strlen(s1));
